infra/requirements.txt
web3==6.15.1
requests==2.32.3
python-dotenv==1.0.1